refactor(engine): convert HistoryLevelSetupCommand to interface
The class is replaced by an interface in the same package so that
ProcessEngineConfiguration signatures remain unchanged. Business
logic moves to DefaultHistoryLevelSetupCommand; static DB helpers
are delegated to HistoryLevelUtils.

Related to #2760
